[BOOK-181] refactor: apis - dto 클래스 도메인 역할에 맞게 재배치
