+++ apis/src/main/kotlin/org/yapp/apis/book/service/UserBookService.kt
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 import org.springframework.validation.annotation.Validated
-import org.yapp.apis.auth.dto.request.UserBooksByIsbnsRequest
+import org.yapp.apis.book.dto.request.UserBooksByIsbnsRequest
 import org.yapp.apis.book.dto.request.UpsertUserBookRequest
 import org.yapp.apis.book.dto.response.UserBookPageResponse
 import org.yapp.apis.book.dto.response.UserBookResponse
